Cleanup: unsigned int to size_t where appropriate

+++ Matrix/BlockCMat.cpp
@@ -21,17 +21,17 @@
 
 #include "BlockCMat.hh"
 
-BlockCMat makeBlockCMatIdentity(unsigned int n, unsigned int mc) {
+BlockCMat makeBlockCMatIdentity(size_t n, size_t mc) {
     BlockCMat foo(n,n,mc);
-    for(unsigned int i=0; i<n; i++)
+    for(size_t i=0; i<n; i++)
         foo(i,i) = CMatrix::identity(mc);
     return foo;
 }
 
-BlockCMat makeBlockCMatRandom(unsigned int n, unsigned int mc) {
+BlockCMat makeBlockCMatRandom(size_t n, size_t mc) {
     BlockCMat foo(n,n,mc);
-    for(unsigned int r=0; r<n; r++)
-        for(unsigned int c=0; c<n; c++)
+    for(size_t r=0; r<n; r++)
+        for(size_t c=0; c<n; c++)
             foo(r,c) = CMatrix::random(mc);
         return foo;
 }
@@ -42,17 +42,17 @@ BlockCMat makeBlockCMatRandom(unsigned int n, unsigned int mc) {
 class Compare_BCM_SVD_singular_values {
 public:
     Compare_BCM_SVD_singular_values(const BlockCMat_SVD* b): B(b) {}
-    bool operator() (unsigned int i, unsigned int j) { return B->getSV(j) < B->getSV(i); }
+    bool operator() (size_t i, size_t j) { return B->getSV(j) < B->getSV(i); }
     const BlockCMat_SVD* B;
 };
 
 
 BlockCMat_SVD::BlockCMat_SVD(const BlockCMat& BC): M(BC.nRows()), N(BC.nCols()), Mc(BC[0].nRows()), Ms(std::min(M,N)), PsI(NULL), PsI_epsilon(0) {
     #ifdef WITH_LAPACKE
-    for(unsigned int i=0; i<Mc/2+1; i++) {
+    for(size_t i=0; i<Mc/2+1; i++) {
         VarMat<lapack_complex_double> dblock(BC.nRows(),BC.nCols());
-        for(unsigned int r=0; r<M; r++) {
-            for(unsigned int c=0; c<N; c++) {
+        for(size_t r=0; r<M; r++) {
+            for(size_t c=0; c<N; c++) {
                 dblock(r,c) = BC(r,c).getKData()[i];
             }
         }
@@ -68,13 +68,13 @@ BlockCMat_SVD::BlockCMat_SVD(const BlockCMat& BC): M(BC.nRows()), N(BC.nCols()),
 BlockCMat_SVD::~BlockCMat_SVD() {
     delete PsI;
     #ifdef WITH_LAPACKE
-    for(unsigned int i=0; i<block_SVDs.size(); i++) delete block_SVDs[i];
+    for(size_t i=0; i<block_SVDs.size(); i++) delete block_SVDs[i];
     #endif
 }
 
-double BlockCMat_SVD::getSV(unsigned int i) const {
+double BlockCMat_SVD::getSV(size_t i) const {
     #ifdef WITH_LAPACKE
-    unsigned int idiag = i/Ms;
+    size_t idiag = i/Ms;
     assert(idiag < Mc);
     if(idiag>=Mc/2+1) idiag = Mc - idiag;
     return block_SVDs[idiag]->singular_values()[i%Ms];
@@ -89,31 +89,31 @@ void BlockCMat_SVD::sort_singular_values() {
     sloc.getData().clear();
     
     // sort internal singular value "ID numbers"
-    for(unsigned int i=0; i<Ms*Mc; i++) sloc.push_back(i);
+    for(size_t i=0; i<Ms*Mc; i++) sloc.push_back(i);
     Compare_BCM_SVD_singular_values CB(this);
     std::sort(sloc.getData().begin(), sloc.getData().end(), CB);
     
     // sorted list of singular values
-    for(unsigned int i=0; i<Ms*Mc; i++) svalues.push_back(getSV(sloc[i]));
+    for(size_t i=0; i<Ms*Mc; i++) svalues.push_back(getSV(sloc[i]));
 }
 
-VarVec<double> BlockCMat_SVD::getRightSVec(unsigned int i) const {
+VarVec<double> BlockCMat_SVD::getRightSVec(size_t i) const {
     i = sloc[i];
     VarVec<double> v;
     #ifdef WITH_LAPACKE
     // check whether this vector belongs to implied complementary set
-    unsigned int idiag = i/Ms;
+    size_t idiag = i/Ms;
     assert(idiag < Mc);
     bool iset = (idiag>Mc/2+1) || (idiag==Mc/2+1 && !(Mc%2));
     if(iset) idiag = Mc - idiag;
     
     VarVec<lapack_complex_double> sv = block_SVDs[idiag]->getRightSVec(i%Ms);
     assert(sv.size()==M);
-    for(unsigned int m=0; m<M; m++) {
+    for(size_t m=0; m<M; m++) {
         CMatrix C(Mc);
         C.getKData()[idiag] = sv[m] * (iset ? complex<double>(0,1) : 1);
         const vector<double>& r = C.getRealData();
-        for(unsigned int mc=0; mc<Mc; mc++) v.push_back(r[mc]);
+        for(size_t mc=0; mc<Mc; mc++) v.push_back(r[mc]);
     }
     #endif
     return v;
@@ -126,10 +126,10 @@ const BlockCMat& BlockCMat_SVD::calc_pseudo_inverse(double epsilon) {
     PsI_epsilon = epsilon;
     
     PsI = new BlockCMat(M, N, Mc);
-    for(unsigned int i=0; i<Mc/2+1; i++) {
+    for(size_t i=0; i<Mc/2+1; i++) {
         const VarMat<lapack_complex_double>& bPsI = block_SVDs[i]->calc_pseudo_inverse(epsilon*svalues[0]);
-        for(unsigned int r=0; r<M; r++) {
-            for(unsigned int c=0; c<N; c++) {
+        for(size_t r=0; r<M; r++) {
+            for(size_t c=0; c<N; c++) {
                 (*PsI)(r,c).getKData()[i] = bPsI(r,c);
             }
         }
@@ -145,7 +145,7 @@ void BlockCMat_SVD::writeToFile(ostream& o) const {
     o.write((char*)&Mc,                 sizeof(Mc));
     #ifdef WITH_LAPACKE
     assert(block_SVDs.size() == Mc/2+1);
-    for(unsigned int i=0; i<block_SVDs.size(); i++)
+    for(size_t i=0; i<block_SVDs.size(); i++)
         block_SVDs[i]->writeToFile(o);
     #endif
     o.write((char*)&PsI,                sizeof(PsI));
@@ -162,7 +162,7 @@ BlockCMat_SVD* BlockCMat_SVD::readFromFile(std::istream& s) {
     s.read((char*)&foo->Mc,             sizeof(foo->Mc));
     foo->Ms = std::min(foo->M,foo->N);
     #ifdef WITH_LAPACKE
-    for(unsigned int i=0; i<foo->Mc/2+1; i++)
+    for(size_t i=0; i<foo->Mc/2+1; i++)
         foo->block_SVDs.push_back( LAPACKE_Matrix_SVD<double,lapack_complex_double>::readFromFile(s) );
     #endif
     s.read((char*)&foo->PsI,            sizeof(foo->PsI));
