+++ Matrix/Permutation.cc
@@ -21,31 +21,31 @@
 
 #include "Permutation.hh"
 
-Permutation::Permutation(unsigned int n): data(vector<unsigned int>(n)) {
-    for(unsigned int i=0; i<n; i++)
+Permutation::Permutation(size_t n): data(vector<size_t>(n)) {
+    for(size_t i=0; i<n; i++)
         data[i] = i;
 }
 
 Permutation& Permutation::nshuffle(int n) {
-    for(unsigned int i=0; i<size()/n; i++) 
+    for(size_t i=0; i<size()/n; i++) 
         for(int j=0; j<n; j++)
             data[j*size()/n+i] = i*n+j;
         return *this;
 }
 
-void Permutation::swap(unsigned int a, unsigned int b) {
+void Permutation::swap(size_t a, size_t b) {
     std::swap(data[a],data[b]);
 }
 
 Permutation Permutation::inverse() const {
     Permutation inv = Permutation(size());
-    for(unsigned int i=0; i<size(); i++) inv[data[i]] = i;
+    for(size_t i=0; i<size(); i++) inv[data[i]] = i;
     return inv;
 }
 
 Permutation& Permutation::invert() {
-    vector<unsigned int> newdat = vector<unsigned int>(size());
-    for(unsigned int i=0; i<size(); i++) newdat[data[i]] = i;
+    vector<size_t> newdat = vector<size_t>(size());
+    for(size_t i=0; i<size(); i++) newdat[data[i]] = i;
     data = newdat;
     return *this;
 }
@@ -53,7 +53,7 @@ Permutation& Permutation::invert() {
 const Permutation Permutation::operator*(const Permutation& p) const {
     assert(p.size() == size());
     Permutation o = Permutation(size());
-    for(unsigned int i=0; i<size(); i++) o[i] = p[data[i]];
+    for(size_t i=0; i<size(); i++) o[i] = p[data[i]];
     return o;
 }
